+++ jest.config.js
@@ -9,9 +9,14 @@ module.exports = {
       useESM: true,
     },
   },
+  moduleDirectories: [
+    "src",
+    "node_modules"
+  ],
   transformIgnorePatterns: ['<rootDir>/node_modules/'],
   modulePathIgnorePatterns: [
     "<rootDir>/tests/kms",
+    "<rootDir>/dist",
   ],
   testTimeout: 50000,
   detectOpenHandles: true