+++ src/site/index.ts
@@ -7,24 +7,24 @@
 //    `npx webpack serve`
 // then navigate to http://locahost:8080
 
+import { CoverCryptHybridDecryption } from "crypto/abe/core/hybrid_crypto/cover_crypt/decryption";
+import { CoverCryptHybridEncryption } from "crypto/abe/core/hybrid_crypto/cover_crypt/encryption";
+import { GpswHybridDecryption } from "crypto/abe/core/hybrid_crypto/gpsw/decryption";
+import { GpswHybridEncryption } from "crypto/abe/core/hybrid_crypto/gpsw/encryption";
+import { EncryptedEntry, WorkerPool } from "crypto/abe/core/hybrid_crypto/worker/worker_pool";
+import { CoverCryptMasterKeyGeneration } from "crypto/abe/core/keygen/cover_crypt_keygen";
+import { GpswMasterKeyGeneration } from "crypto/abe/core/keygen/gpsw_crypt_keygen";
+import { logger } from "utils/logger";
+import { hexDecode, hexEncode } from "utils/utils";
+import { EncryptionDecryptionDemo } from "../../tests/crypto/abe/common/demo_hybrid_crypto";
+import { CoverCryptDemoKeys } from "../../tests/crypto/abe/core/cover_crypt/demo_keys";
+import { GpswDemoKeys } from "../../tests/crypto/abe/core/gpsw/demo_keys";
+import { generateCoverCryptKeys } from "../../tests/crypto/sse/findex/implementations/common/cover_crypt_keys";
+import { masterKeysFindex } from "../../tests/crypto/sse/findex/implementations/common/keys";
+import { Users, User } from "../../tests/crypto/sse/findex/implementations/common/users";
+import { CloudproofDemoPostgRest } from "../../tests/crypto/sse/findex/implementations/postgrest/cloudproof";
+import { PostgRestDB } from "../../tests/crypto/sse/findex/implementations/postgrest/db";
 
-import { CoverCryptHybridDecryption } from "../crypto/abe/hybrid_crypto/cover_crypt/decryption"
-import { CoverCryptHybridEncryption } from "../crypto/abe/hybrid_crypto/cover_crypt/encryption"
-import { GpswHybridDecryption } from "../crypto/abe/hybrid_crypto/gpsw/decryption"
-import { GpswHybridEncryption } from "../crypto/abe/hybrid_crypto/gpsw/encryption"
-import { EncryptedEntry, WorkerPool } from "../crypto/abe/hybrid_crypto/worker/worker_pool"
-import { CoverCryptMasterKeyGeneration } from "../crypto/abe/keygen/cover_crypt/cover_crypt_keygen"
-import { GpswMasterKeyGeneration } from "../crypto/abe/keygen/gpsw/gpsw_crypt_keygen"
-import { CoverCryptDemoKeys } from "../demos/abe/cover_crypt/demo_keys"
-import { EncryptionDecryptionDemo } from "../demos/abe/demo_hybrid_crypto"
-import { GpswDemoKeys } from "../demos/abe/gpsw/demo_keys"
-import { generateCoverCryptKeys } from "../demos/findex/cover_crypt_keys"
-import { masterKeysFindex } from "../demos/findex/keys"
-import { CloudproofDemoPostgRest } from "../demos/findex/postgrest/cloudproof"
-import { PostgRestDB } from "../demos/findex/postgrest/db"
-import { User, Users } from "../demos/findex/users"
-import { logger } from "./../utils/logger"
-import { hexDecode, hexEncode } from "./../utils/utils"
 
 const FINDEX_DEMO = new CloudproofDemoPostgRest(new PostgRestDB());
 const COVER_CRYPT_KEYS = generateCoverCryptKeys();