+++ tests/kms/certificateRequest.test.ts
@@ -1,5 +1,5 @@
-import { CertificateRequest } from "../../src/kms/objects/CertificateRequest"
-import { CertificateRequestType } from "../../src/kms/types/CertificateRequestType"
+import { CertificateRequest } from "kms/objects/CertificateRequest"
+import { CertificateRequestType } from "kms/types/CertificateRequestType"
 
 const cr = new CertificateRequest(CertificateRequestType.CRMF, [8, 16, 64, 256])